fix doc strings, move cdd_api
