+++ asapdiscovery-data/asapdiscovery/data/tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from asapdiscovery.data.cdd_api.cdd_api import CDDAPI
+from asapdiscovery.data.cdd_api import CDDAPI
 from asapdiscovery.data.schema_v2.complex import Complex, PreppedComplex
 from asapdiscovery.data.schema_v2.ligand import Ligand
 from asapdiscovery.data.services_config import CDDSettings
